third_party/WebKit/LayoutTests/fast/alignment/ensure-flexbox-compatibility-with-initial-values-expected.txt

Test to verify initial values of alignment properties are backward-comaptible with flexbox implementation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS CSS.supports('align-items', 'start') is true
PASS CSS.supports('align-self', 'start') is true
PASS CSS.supports('align-content', 'normal') is true
PASS CSS.supports('justify-content', 'normal') is true
PASS CSS.supports('align-items', 'start') is true
PASS CSS.supports('align-self', 'start') is true
PASS CSS.supports('align-content', 'normal') is true
PASS CSS.supports('justify-content', 'normal') is true
PASS CSS.supports('align-items', 'stretch') is true
PASS CSS.supports('align-self', 'start') is true
PASS CSS.supports('align-content', 'normal') is true
PASS CSS.supports('justify-content', 'normal') is true
PASS CSS.supports('align-items', 'start') is true
PASS CSS.supports('align-self', 'stretch') is true
PASS CSS.supports('align-content', 'normal') is true
PASS CSS.supports('justify-content', 'normal') is true
PASS successfullyParsed is true

TEST COMPLETE
